% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyDescribe.R
\docType{package}
\name{EasyDescribe-package}
\alias{EasyDescribe-package}
\alias{EasyDescribe}
\title{A Convenient Way of Descriptive Statistics}
\description{
EasyDescribe provide a convenient way of descriptive statistics.
}
