% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2D.R
\docType{data}
\name{T2D}
\alias{T2D}
\title{A data for 20 diabetes patients}
\format{
A data.frame contains 20 obs of 8 variables. The variables are:
\describe{
  \item{ID}{The ID of these 20 people.}
  \item{gender}{A character ("F" and "M").}
  \item{age}{A numeric.}
  \item{education}{A ordered factor ("1"<"2"<"3"<"4").}
  \item{marriage}{A logical.}
  \item{smoke}{A factor (0: "never smoker", 1: "current smoker", 2: "ex-smoker").}
  \item{glucose}{A numeric.}
  \item{diabetes}{A factor (0: "normal people", 1: "patient").}
}
}
\usage{
T2D
}
\description{
A data for 20 diabetes patients. The data were fabricated.
}
\keyword{datasets}
